\name{load.sum}
\alias{load.sum}
\title{
Appropriate Summary Statistics for Factors, Ordered Factors, and Numeric Variables
}
\description{
Returns the medians of a vector of ordered factor variables, the modes of a vector of unordered factor variables, and the means of a vector of numeric variables.  
}
\usage{
load.sum(vec)
}
\arguments{
  \item{vec}{
A vector of ordered factor variables, a vector of unordered factor variables, or a vector of numeric variables.
}
}
\value{
\code{load.sum} returns the medians of a vector of ordered factor variables, the mode of a vector of unordered factor variables, and the mean of a vector of numeric variables.
}
\author{
Michael Lipsitz, Alexandre Belloni, Victor Chernozhukov, Ivan Fernandez-Val
}

\seealso{
\code{\link{npqr}}
}