\name{msqrt}
\alias{msqrt}
\title{
Square Root of Matrix by Spectral Decomposition
}
\description{
Obtains the square root of a symmetric matrix by spectral decomposition.
}
\usage{
msqrt(a)
}
\arguments{
  \item{a}{
a matrix
}
}
\value{
\code{msqrt} returns the square root of a symmetric matrix, obtained via spectral decomposition
}
\author{
Michael Lipsitz, Alexandre Belloni, Victor Chernozhukov, Ivan Fernandez-Val
}

\seealso{
\code{\link{npqr}}
}